Fixed the issue of not being able to scroll all the way to the end when soft keyboard was shown.
Description:
The issue was fixed using `android:fitsSystemWindows="true"` with `android:windowSoftInputMode="adjustResize"`. Added `android:fitsSystemWindows="true"` to our `NestedScrollView` in our activity_main.xml layout file and `android:windowSoftInputMode="adjustResize"` in our manifest file within the <activity> tag of MainActivity.

How did it fix the issue?

The `android:windowSoftInputMode="adjustResize"` attribute in your Android Manifest file tells the system to resize your activity’s main window to make room for the soft keyboard.

Normally, when the soft keyboard is displayed, it won’t resize the window but rather ‘push’ it up off the bottom of the screen. This means that if our `NestedScrollView` is taller than the remaining space above the keyboard, the bottom part of it will be cut off.

By setting `android:windowSoftInputMode="adjustResize"`, we're telling the system to resize our activity’s main window to be smaller to make room for the soft keyboard. This means that the `NestedScrollView` will also be resized and we’ll be able to scroll to the bottom of it even when the keyboard is displayed.

Now let's understand `android:fitsSystemWindows="true"`. Imagine our phone screen as a big picture frame. Normally, everything we put in this frame can go all the way to the edges of the frame. This includes our app’s content, but also the system’s status bar at the top and the navigation bar at the bottom.

Now, if we don’t want our app’s content to go under these bars and be partially hidden by them, we can use `android:fitsSystemWindows="true"`. This is like telling our app “Hey, don’t put your stuff under the system’s bars. Make sure everything fits within the remaining space.”

So, `android:fitsSystemWindows="true"` is like a rule that helps your app play nicely with the system’s bars. It doesn’t directly deal with the soft keyboard or change when the keyboard is on or off. But it helps make sure your app’s content isn’t hidden behind system windows, including the soft keyboard when it’s shown.