Added functionality to show error message on text fields when text fields are empty.
I've added two functions:

1. showErrorWhenTextFieldIsEmpty:

This function simply encapsulates the following and core functions for the functionality.

2. showErrorWhenTextFieldIsEmptyAfterFocus:

This function shows error message when text field is left empty after being focused. It's because, by default, the text fields are empty. And when user comes to enter a value and backs off, the text fields still remains empty which not expected behavior so it cautions the user.

The text fields comes empty, but it's not expected to leave empty.

3. showErrorWhenTextFieldIsEmptyAfterTextChange:

This function shows error message when the user leaves the text field empty after clearing the previous data.