Solved the issue of overlapping between the clear text button and the error exclamation mark at the end of the current elec meter reading text field.
Performed `setError()` on `TextInputLayout` instead of `TextInputEditText`.

When we perform `setError()` on `TextInputEditText`, the error message is show only when the text field is focuses else there will be only a red exclamation mark at the end of the text field, and it also overlaps if there is trailing icon or button like "clear_text" button. The error message is also shown as dialog linked with the exclamation.

On the other hand, if we perform `setError()` on the `TextInputLayout`, the error message is shown below the text field which is the default Material behavior and the error message is shown regardless of whether the text field is focused or not. It also doesn't overlap the trailing icon or button.