Removed `ScrollBehavior` from `CenterAlignedTopAppBar` to fix the unnecessary scrolling of top app bar.
The issue with the unnecessary scrolling of the TopAppBar in the InvoicePage was due to the use of the TopAppBarDefaults.enterAlwaysScrollBehavior() scroll behavior. This behavior causes the app bar to react to scroll gestures, even when the content does not exceed the screen height, leading to an undesirable user experience. The solution was to remove the scroll behavior by setting it to null, ensuring that the app bar remains fixed when the content does not require scrolling. This change effectively prevents the app bar from moving unnecessarily, providing a more stable and expected interface for users.